+++ CHANGELOG.rst
@@ -7,7 +7,8 @@ Current
 - Handle callable on API infos
 - Drop/merge flask_restful `flask_restful.RequestParser`
 - Handle :class:`~flask_restplus.reqparse.RequestParser` into :meth:`~flask_restplus.Api.expect` decorator
-- Handle schema for :class:`~flask_restplus.inputs` parsers
+- Handle schema for :mod:`~flask_restplus.inputs` parsers
+- Added :class:`~flask_restplus.inputs.email` and :func:`~flask_restplus.inputs.ip`
 
 0.8.5 (2015-12-12)
 ------------------